requirements.txt
flask
langchain
faiss-cpu
transformers
sentence-transformers
gitpython
torch
langchain-community
langchain-huggingface
